requirements.txt
PyQt6
yt-dlp
Pillow
requests
packaging
qtawesome
psutil